remove supervisord

+++ image/prepare.sh
@@ -15,7 +15,7 @@ check_env () {
     fi
 }
 
-if [ "$1" == "/usr/bin/supervisord" ] && [ "$EUID" -eq "0" ]; then
+prepare() {
     # Check environment variables
     echo "Checking environment variables..."
     export PUID=${PUID:-33}
@@ -46,39 +46,54 @@ if [ "$1" == "/usr/bin/supervisord" ] && [ "$EUID" -eq "0" ]; then
     if [ -c "/dev/dri/renderD128" ]; then
         chgrp www-data /dev/dri/renderD128
     fi
+}
 
-    for dir in /var/www/html /var/www/cache; do
-        echo "Checking permissions for ${dir}..."
-        find "${dir}" ! -user www-data -or ! -group www-data \
-            -exec chown --no-dereference www-data:www-data {} \; \
-            -print
-    done
-    echo "All things prepared."
+if [ "$EUID" -eq "0" ]; then
+    case "$1" in
+        apache2-foreground)
+            prepare
+            for dir in /var/www/html /var/www/cache; do
+                echo "Checking permissions for ${dir}..."
+                find "${dir}" ! -user www-data -or ! -group www-data \
+                    -exec chown --no-dereference www-data:www-data {} \; \
+                    -print
+            done
+            # Run /entrypoint.sh logic (install/upgrade/...)
+            export NEXTCLOUD_UPDATE=1
+            export NEXTCLOUD_INIT_HTACCESS=1
+            /entrypoint.sh true
 
-    # Run /entrypoint.sh logic (install/upgrade/...)
-    export NEXTCLOUD_UPDATE=1
-    export NEXTCLOUD_INIT_HTACCESS=1
-    /entrypoint.sh true
+            # Install notify_push
+            rsync -a --delete --chown www-data:www-data \
+                /usr/src/nextcloud/apps/notify_push/ \
+                /var/www/html/apps/notify_push/
+            occ app:enable notify_push
 
-    # Prepare notify_push
-    rsync -a --delete --chown www-data:www-data \
-        /usr/src/nextcloud/apps/notify_push/ \
-        /var/www/html/apps/notify_push/
-    occ app:enable notify_push
-
-    # Config manually in case reverse proxy not ready
-    if [ "$(occ config:app:get notify_push base_endpoint)" != "${OVERWRITECLIURL}/push" ]; then
-        # Random value [0, 2**30-1]
-        occ config:app:set notify_push cookie --value "$(($RANDOM*2**15+$RANDOM))"
-        occ config:app:set notify_push base_endpoint --value "${OVERWRITECLIURL}/push"
-    fi
-
-    # ENV
-    export DATABASE_URL=postgres://${POSTGRES_USER}:${POSTGRES_PASSWORD}@${POSTGRES_HOST}/${POSTGRES_DB}
-    export DATABASE_PREFIX=oc_
-    export REDIS_URL=redis://${REDIS_HOST}
-    export NEXTCLOUD_URL=http://${NEXTCLOUD_HOST}
-
+            # Config manually in case reverse proxy not ready
+            if [ "$(occ config:app:get notify_push base_endpoint)" != "${OVERWRITECLIURL}/push" ]; then
+                # Random value [0, 2**30-1]
+                occ config:app:set notify_push cookie --value "$(($RANDOM*2**15+$RANDOM))"
+                occ config:app:set notify_push base_endpoint --value "${OVERWRITECLIURL}/push"
+            fi
+            echo Starting: "$@"
+            exec "$@"
+        ;;
+        supercronic)
+            prepare
+            echo Starting: "$@"
+            exec /su.py www-data "$@"
+        ;;
+        notify_push)
+            prepare
+            export DATABASE_URL=postgres://${POSTGRES_USER}:${POSTGRES_PASSWORD}@${POSTGRES_HOST}/${POSTGRES_DB}
+            export DATABASE_PREFIX=oc_
+            export REDIS_URL=redis://${REDIS_HOST}
+            export NEXTCLOUD_URL=http://${NEXTCLOUD_HOST}
+            echo Starting: "$@"
+            sleep 5
+            exec /su.py www-data "$@"
+        ;;
+    esac
 fi
 
 exec "$@"
